fix 64-bit issues, clear signing settings

+++ game/Builder.cpp
@@ -586,7 +586,7 @@ void MobileUnitBuildForm::SetOwner(BuilderGob *pbldr)
 
 void MobileUnitBuildForm::UpdateUnitInfo(ListItem *pli)
 {
-	MobileUnitConsts *pmuntc = (MobileUnitConsts *)gapuntc[(int)pli->pvData];
+	MobileUnitConsts *pmuntc = (MobileUnitConsts *)gapuntc[(long)pli->pvData];
 
 	// Update Cost
